+++ packages/experiments/src/components/Stack/HorizontalStack/HorizontalStack.styles.ts
@@ -16,8 +16,8 @@ export const styles = (props: IThemedProps<IHorizontalStackProps>): IHorizontalS
   const hGap = parseGap(gap, theme);
   const vGap = parseGap(vertGap, theme);
 
-  const horizontalMargin = -0.5 * hGap.value;
-  const verticalMargin = -0.5 * vGap.value;
+  const horizontalMargin = `${-0.5 * hGap.value}${hGap.unit}`;
+  const verticalMargin = `${-0.5 * vGap.value}${vGap.unit}`;
 
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
 
@@ -30,27 +30,31 @@ export const styles = (props: IThemedProps<IHorizontalStackProps>): IHorizontalS
           maxHeight,
           width: fillHorizontal ? '100%' : 'auto',
           height: fillVertical ? '100%' : 'auto',
-          overflow: 'visible',
-          display: 'block',
-
-          // necessary in order to prevent collapsing margins
-          paddingTop: 1
+          overflow: 'visible'
         },
-        className
+        className,
+        {
+          // not allowed to be overridden by className
+          // since this is necessary in order to prevent collapsing margins
+          display: 'inline-block'
+        }
       ],
 
       inner: [
         classNames.inner,
         {
           flexWrap: 'wrap',
-          margin: `${verticalMargin}${vGap.unit} ${horizontalMargin}${hGap.unit}`,
-
-          // account for the extra 1px padding at the top of the root
-          marginTop: `calc(${verticalMargin}${vGap.unit} - 1px)`,
-
+          marginLeft: horizontalMargin,
+          marginRight: horizontalMargin,
+          marginTop: verticalMargin,
+          marginBottom: verticalMargin,
           overflow: 'visible',
-          width: fillHorizontal ? `calc(100% + ${hGap.value}${hGap.unit})` : 'auto',
-          height: fillVertical ? `calc(100% + ${hGap.value}${hGap.unit})` : 'auto'
+          maxWidth: '100vw',
+
+          // avoid unnecessary calc() calls if vertical gap is 0
+          height: fillVertical ?
+            (vGap.value === 0 ? '100%' : `calc(100% + ${vGap.value}${vGap.unit})`)
+            : 'auto'
         }
       ]
     } as IHorizontalStackStyles;