+++ packages/merge-styles/src/styleToClassName.ts
@@ -68,13 +68,15 @@ function extractRules(args: IStyle[], rules: IRuleSet = { __order: [] }, current
             }
           }
         } else {
-          // Else, add the rule to the currentSelector.
-          if (prop === 'margin' || prop === 'padding') {
-            // tslint:disable-next-line:no-any
-            expandQuads(currentRules, prop, (arg as any)[prop]);
-          } else {
-            // tslint:disable-next-line:no-any
-            (currentRules as any)[prop] = (arg as any)[prop] as any;
+          if ((arg as any)[prop] !== undefined) {
+            // Else, add the rule to the currentSelector.
+            if (prop === 'margin' || prop === 'padding') {
+              // tslint:disable-next-line:no-any
+              expandQuads(currentRules, prop, (arg as any)[prop]);
+            } else {
+              // tslint:disable-next-line:no-any
+              (currentRules as any)[prop] = (arg as any)[prop] as any;
+            }
           }
         }
       }